multiview

+++ mask/get_raw_mask.py
@@ -190,25 +190,30 @@ if __name__ == "__main__":
     print("Generating masks...")
     for image_name in tqdm(sorted(os.listdir(image_folder))):
         image_path = os.path.join(image_folder, image_name)
-        if args.seg_method == "sam":
-            image = cv2.imread(image_path)
-            mask = get_sam_mask(seg_model, image, config["confidence_threshold"])
-        elif args.seg_method == "entityseg":
-            from detectron2.data.detection_utils import read_image
-            image = read_image(image_path, format="BGR")
-            mask = get_entityseg_mask(seg_model, image, config["confidence_threshold"])
-        # elif args.seg_method == "panopticseg":
-        #     from detectron2.data.detection_utils import read_image
-        #     image = read_image(image_path, format="BGR")
-        #     mask = get_panopticseg_mask(seg_model, image, config["confidence_threshold"])
-        else:
-            raise NotImplementedError
-
-        # mask_path = os.path.join(output_folder, image_name.replace(".jpg", ".png"))
         mask_path = os.path.join(output_folder, image_name.split(".")[0] + ".png")
-        cv2.imwrite(mask_path, mask)
+        
+        if not os.path.exists(mask_path):
+            if args.seg_method == "sam":
+                image = cv2.imread(image_path)
+                mask = get_sam_mask(seg_model, image, config["confidence_threshold"])
+            elif args.seg_method == "entityseg":
+                from detectron2.data.detection_utils import read_image
+                image = read_image(image_path, format="BGR")
+                mask = get_entityseg_mask(seg_model, image, config["confidence_threshold"])
+            # elif args.seg_method == "panopticseg":
+            #     from detectron2.data.detection_utils import read_image
+            #     image = read_image(image_path, format="BGR")
+            #     mask = get_panopticseg_mask(seg_model, image, config["confidence_threshold"])
+            else:
+                raise NotImplementedError
 
-        if args.visualize:
-            vis_mask = visualize_mask(mask)
-            vis_mask_path = os.path.join(vis_output_folder, image_name.split(".")[0] + ".png")
-            cv2.imwrite(vis_mask_path, vis_mask)
+            # mask_path = os.path.join(output_folder, image_name.replace(".jpg", ".png"))
+            # mask_path = os.path.join(output_folder, image_name.split(".")[0] + ".png")
+            cv2.imwrite(mask_path, mask)
+
+            if args.visualize:
+                vis_mask = visualize_mask(mask)
+                vis_mask_path = os.path.join(vis_output_folder, image_name.split(".")[0] + ".png")
+                cv2.imwrite(vis_mask_path, vis_mask)
+        else:
+            print(f"{mask_path} already existed..")
